refactor(output): Add components barrel for cleaner exports
