+++ src/cli/output/index.ts
@@ -1,7 +1,4 @@
-export type { FileError } from "./components/errors";
-export { ProgressBar } from "./components/progressBar";
-export type { Warning } from "./components/warnings";
+export * from "./components";
 export * from "./formatter";
-export { OutputRenderer } from "./renderer";
-export type { RendererOptions, SuccessData } from "./renderer";
+export * from "./renderer";
 export * from "./styles";
